+++ src/Exceptions/InvalidInvoiceItemException.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the InvalidInvoiceItemException class.
+ *
+ * @copyright   Copyright (c) 2022 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2022-10-06
+ *
+ */
+
+namespace Konekt\Factureaza\Exceptions;
+
+class InvalidInvoiceItemException extends ValidationException
+{
+}