+++ src/Exceptions/ValidationException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ValidationException class.
+ *
+ * @copyright   Copyright (c) 2022 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2022-10-06
+ *
+ */
+
+namespace Konekt\Factureaza\Exceptions;
+
+use Exception;
+
+class ValidationException extends Exception
+{
+}