+++ src/Models/Client.php
@@ -3,32 +3,22 @@
 declare(strict_types=1);
 
 /**
- * Contains the Invoice class.
+ * Contains the Client class.
  *
  * @copyright   Copyright (c) 2022 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2022-09-17
+ * @since       2022-10-06
  *
  */
 
 namespace Konekt\Factureaza\Models;
 
-use Carbon\CarbonImmutable;
 use Konekt\Factureaza\Contracts\Resource;
 
-class Invoice implements Resource
+class Client implements Resource
 {
     use HasDynamicAttributeConstructor;
-    use HasId;
-    use HasTimestamps;
-
-    public readonly CarbonImmutable $documentDate;
-
-    public readonly string $clientId;
-
-    /** @var InvoiceItem[] */
-    public readonly array $items;
 
     public static function attributeMap(): array
     {