+++ Changelog.md
@@ -4,3 +4,6 @@
 ##### 2022-XX-YY
 
 - Initial release
+- Sandbox or productions accounts can be used
+- My Account can be queried
+- Invoices can be created